requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
ape-alchemy==0.6.4
ape-ens==0.6.1
ape-etherscan==0.6.10
ape-foundry==0.6.15
ape-ganache==0.6.8
ape-hardhat==0.6.12
ape-infura==0.6.3
ape-solidity==0.6.8
ape-template==0.6.1
ape-tokens==0.6.2
ape-vyper==0.6.10
appnope==0.1.3
arrow==1.2.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
base58==1.0.3
binaryornot==0.4.4
bitarray==2.8.1
cached-property==1.5.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chompjs==1.2.2
click==8.1.7
commonmark==0.9.1
cookiecutter==2.1.1
cryptography==41.0.3
cytoolz==0.12.2
dataclassy==0.11.1
decorator==5.1.1
Deprecated==1.2.14
eip712==0.2.1
eth-abi==4.2.0
eth-account==0.8.0
eth-ape==0.6.19
eth-bloom==2.0.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-tester==0.9.1b1
eth-typing==3.4.0
eth-utils==2.2.0
ethpm-types==0.5.4
evm-trace==0.1.0a24
exceptiongroup==1.1.3
executing==1.2.0
flake8==6.1.0
frozenlist==1.4.0
greenlet==2.0.2
hexbytes==0.3.1
idna==3.4
ijson==3.2.3
importlib-metadata==6.8.0
iniconfig==2.0.0
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lazyasd==0.1.4
lru-dict==1.2.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
morphys==1.0
msgspec==0.18.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.25.2
packaging==23.1
pandas==1.5.3
parsimonious==0.9.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.3.0
prompt-toolkit==3.0.39
protobuf==4.24.2
ptyprocess==0.7.0
pure-eval==0.2.2
py-cid==0.3.0
py-ecc==6.0.0
py-evm==0.7.0a4
py-geth==3.13.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==0.2.3
py-solc-x==1.1.1
pycodestyle==2.11.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.12
pyethash==0.1.27
pyflakes==3.1.0
PyGithub==1.59.1
Pygments==2.16.1
PyJWT==2.8.0
PyNaCl==1.5.0
pytest==7.4.1
python-baseconv==1.2.2
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
pyunormalize==15.0.0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==12.6.0
rlp==3.0.0
rpds-py==0.10.2
safe-pysha3==1.0.4
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.20
stack-data==0.6.2
text-unidecode==1.3
tokenlists==0.1.4
tomli==2.0.1
toolz==0.12.0
tqdm==4.66.1
traitlets==5.9.0
trie==2.1.1
typing_extensions==4.7.1
urllib3==2.0.4
varint==1.0.2
vvm==0.1.0
watchdog==3.0.0
wcwidth==0.2.6
web3==6.9.0
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.2
